+++ src/aiidalab_qe/setup/codes.py
@@ -1,13 +1,13 @@
+import subprocess
 from pathlib import Path
 from shutil import which
-from subprocess import CalledProcessError, run
-
 from filelock import FileLock, Timeout
 
 from aiida.common.exceptions import NotExistent
 from aiida.orm import load_code
 
-FN_LOCKFILE = Path.home().joinpath(".install-qe-on-localhost.lock")
+FN_INSTALL_LOCKFILE = Path.home().joinpath(".install-qe-on-localhost.lock")
+FN_SETUP_LOCKFILE = Path.home().joinpath(".setup-qe-on-localhost.lock")
 FN_DO_NOT_SETUP = Path.cwd().joinpath(".do-not-setup-on-localhost")
 
 QE_VERSION = "7.2"
@@ -43,11 +43,19 @@ CODE_NAMES = (
 
 
 def qe_installed():
-    return get_qe_env().exists()
+    env_exist = get_qe_env().exists()
+    proc = subprocess.run(
+        ["conda", "list", "-n", f"{get_qe_env().name}", "qe"],
+        check=True,
+        capture_output=True,
+    )
+
+    # XXX: "qe" in check is not future proof if there are similar packages such as qe-tool, better solution?? JSON output??
+    return env_exist and "qe" in str(proc.stdout)
 
 
 def install_qe():
-    run(
+    subprocess.run(
         [
             "conda",
             "create",
@@ -64,17 +72,17 @@ def install_qe():
     )
 
 
-def _code_is_setup(name):
+def _code_is_setup(name, computer):
     try:
-        load_code(f"{name}-{QE_VERSION}@localhost")
+        load_code(f"{name}-{QE_VERSION}@{computer}")
     except NotExistent:
         return False
     else:
         return True
 
 
-def codes_are_setup():
-    return all(_code_is_setup(code_name) for code_name in CODE_NAMES)
+def codes_are_setup(computer):
+    return all(_code_is_setup(code_name, computer) for code_name in CODE_NAMES)
 
 
 def _generate_header_to_setup_code():
@@ -89,13 +97,13 @@ load_profile()
     return header_code
 
 
-def _generate_string_to_setup_code(code_name, computer_name="localhost"):
+def _generate_string_to_setup_code(code_name, computer):
     """Generate the Python string to setup an AiiDA code for a given computer.
 
     Tries to load an existing code and if not existent,
     generates Python code to create and store a new code setup."""
     try:
-        load_code(f"{code_name}-{QE_VERSION}@{computer_name}")
+        load_code(f"{code_name}-{QE_VERSION}@{computer}")
     except NotExistent:
         label = f"{code_name}-{QE_VERSION}"
         description = f"{code_name}.x ({QE_VERSION}) setup by AiiDAlab."
@@ -114,7 +122,7 @@ code = InstalledCode(computer=computer,
 
 code.store()
 """.format(  # noqa: UP032
-            computer_name,
+            computer,
             label,
             description,
             filepath_executable,
@@ -127,77 +135,110 @@ code.store()
         return ""
 
 
-def setup_codes():
+def setup_codes(computer):
     python_code = _generate_header_to_setup_code()
     for code_name in CODE_NAMES:
-        python_code += _generate_string_to_setup_code(code_name)
+        python_code += _generate_string_to_setup_code(code_name, computer)
     try:
-        run(["python", "-c", python_code], capture_output=True, check=True)
-    except CalledProcessError as error:
-        raise RuntimeError(f"Failed to setup codes: {error}") from None
+        subprocess.run(["python", "-c", python_code], capture_output=True, check=True)
+    except subprocess.CalledProcessError as err:
+        raise RuntimeError(
+            f"Failed to setup codes, exit_code={err.returncode}, {err.stderr}"
+        ) from None
 
 
-def install(force=False):
+def install_and_setup(computer="localhost", force=False):
     """Install Quantum ESPRESSO and the corresponding AiiDA codes.
 
     Args:
         force: Ignore previously failed attempts and install anyways.
+        computer: computer label in AiiDA where the code is setup for
     """
     # Check for "do not install file" and skip actual check. The purpose of
     # this file is to not re-try this process on every app start in case that
     # there are issues.
+    # XXX: use filelock to control `FN_DO_NOT_SETUP` as well
     if not force and FN_DO_NOT_SETUP.exists():
         raise RuntimeError("Installation failed in previous attempt.")
 
+    yield from _install()
+    yield from _setup(computer)
+
+
+def _install():
+    """Install Quantum ESPRESSO."""
     yield "Checking installation status..."
 
     conda_installed = which("conda")
     try:
-        with FileLock(FN_LOCKFILE, timeout=5):
-            # We assume that if the codes are already setup, everything is in
-            # order. Only if they are not present, should we take action,
-            # however we only do so if the environment has a conda binary
-            # present (`which conda`). If that is not the case then we assume
-            # that this is a custom user environment in which case we also take
-            # no further action.
-            if codes_are_setup():
-                return  # Already setup
-
+        with FileLock(FN_INSTALL_LOCKFILE, timeout=5):
             if not conda_installed:
                 raise RuntimeError(
                     "Unable to automatically install Quantum ESPRESSO, conda "
                     "is not available."
                 )
 
+            if qe_installed():
+                return
+
+            # Install Quantum ESPRESSO.
+            yield "Installing QE..."
+            try:
+                install_qe()
+            except subprocess.CalledProcessError as error:
+                raise RuntimeError(
+                    f"Failed to create conda environment: {error}"
+                ) from None
+
+    except Timeout:
+        # Assume that the installation was triggered by a different process.
+        yield "Installation was already started, waiting for it to finish..."
+        with FileLock(FN_INSTALL_LOCKFILE, timeout=120):
             if not qe_installed():
-                # First, install Quantum ESPRESSO.
-                yield "Installing QE..."
-                try:
-                    install_qe()
-                except CalledProcessError as error:
-                    raise RuntimeError(
-                        f"Failed to create conda environment: {error}"
-                    ) from None
+                raise RuntimeError(
+                    "Installation process did not finish in the expected time."
+                ) from None
+
+
+def _setup(computer):
+    """Setup the corresponding AiiDA codes after QE installation."""
+    yield "Checking setup status..."
+
+    try:
+        with FileLock(FN_SETUP_LOCKFILE, timeout=5):
+            # We assume that if the codes are already setup, everything is in
+            # order. Only if they are not present, should we take action,
+            # however we only do so if the environment has a conda binary
+            # present (`which conda`). If that is not the case then we assume
+            # that this is a custom user environment in which case we also take
+            # no further action.
+            if codes_are_setup(computer=computer):
+                return  # Already setup
 
             # After installing QE, we install the corresponding
             # AiiDA codes:
             python_code = _generate_header_to_setup_code()
             for code_name in CODE_NAMES:
-                if not _code_is_setup(code_name):
-                    yield f"Preparing setup script for ({code_name})..."
-                    code_string = _generate_string_to_setup_code(code_name)
+                if not _code_is_setup(code_name, computer=computer):
+                    yield f"Preparing setup script for ({code_name}) on ({computer})..."
+                    code_string = _generate_string_to_setup_code(code_name, computer)
                     python_code += code_string
             try:
                 yield "Setting up all codes..."
-                run(["python", "-c", python_code], capture_output=True, check=True)
-            except CalledProcessError as error:
-                raise RuntimeError(f"Failed to setup codes: {error}") from None
+                subprocess.run(
+                    ["python", "-c", python_code], capture_output=True, check=True
+                )
+            except subprocess.CalledProcessError as err:
+                raise RuntimeError(
+                    f"Failed to setup codes, exit_code={err.returncode}, {err.stderr}"
+                ) from None
 
     except Timeout:
         # Assume that the installation was triggered by a different process.
         yield "Installation was already started, waiting for it to finish..."
-        with FileLock(FN_LOCKFILE, timeout=120):
-            if not codes_are_setup():
+        with FileLock(FN_SETUP_LOCKFILE, timeout=120):
+            if not codes_are_setup(computer=computer):
                 raise RuntimeError(
                     "Installation process did not finish in the expected time."
                 ) from None
+
